apply changes prompt extracted
